// types/types.d.ts
import { TextInputProps, TouchableOpacityProps } from "react-native";


declare interface ButtonProps extends TouchableOpacityProps {
    title?: string;
    bgVariant?: "primary" | "secondary" | "danger" | "outline" | "success";
    textVariant?: "primary" | "default" | "secondary" | "danger" | "success";
    IconLeft?: React.ComponentType<any>;
    IconRight?: React.ComponentType<any>;
    className?: string;
    disabled?:boolean,
    loading?:boolean,
}

type SelectedCategory = {
    name: string;
    subCategories: {
        name: string;
    }[];
};

export type SelectedCategoriesRequest = {
    selectedCategories: SelectedCategory[];
};